# src/main/resources/static/openapi/banking-openapi.yaml
openapi: 3.0.3
info:
  title: Brite Banking API
  version: 1.0.0
  description: |
    REST API for the `org.bee.banking` module of the SBProjects Spring Boot application:
    account registration, lookup, deposit/withdrawal, closing (single and bulk), paginated
    account-status search, bank office/ATM location search, bank statements, plus the module's async-notification and
    Resilience4j payment demo endpoints.

    ## Gateway behavior (applies to every endpoint in this document)

    Every path below sits behind two servlet filters (`org.bee.banking.gateway`), which run
    *before* the controllers:

    1. **`BusinessTransactionIdFilter`** generates a UUID per request and returns it in the
       `X-BTID` response header. The same id is stamped on every server log line for that
       request, so quote it when reporting a problem.
    2. **`BankingRateLimitFilter`** requires the customer-id header (default `X-Customer-Id`,
       configurable via `banking.rate-limit.customer-header-name`) and caps each customer at
       `banking.rate-limit.requests-per-day` requests/day (default 1000). A missing/blank
       header is rejected with `400`; an exhausted quota with `429`. Both are **plain-text**
       bodies written by the filter itself, not the JSON/text bodies produced by the
       controllers' exception handler. Every non-rejected response carries
       `X-RateLimit-Limit` and `X-RateLimit-Remaining`.

...
    authentication - the API has no auth layer.

...

...

...

...
servers:
...
security:
  - CustomerId: []
paths:
...
  /v1/api/locations:
    get:
      tags: [Locations]
      operationId: listBankLocations
      summary: Search bank offices and ATMs (paginated)
      description: |
        Returns bank locations filtered by any combination of the optional filters, paginated
        and sorted with real SQL (Spring Data `Specification`). Not cached.

        **`type` matches by capability**, not by the stored value: `OFFICE` returns offices
        *and* office+ATM branches (`BOTH`), `ATM` returns ATMs *and* `BOTH` branches, and
        `BOTH` returns only branches that have both. `city`, `state` and `zip` are
        case-insensitive exact matches; `service` matches locations that serve that operation.

        Office hours (`opensAt`/`closesAt`) and `phoneNumber` belong to the office, so they are
        `null` for `ATM` locations.
      parameters:
        - name: type
          in: query
          required: false
          schema: {$ref: '#/components/schemas/LocationType'}
        - name: city
          in: query
          required: false
          schema: {type: string}
          example: Chicago
        - name: state
          in: query
          required: false
          description: Two-letter state code (case-insensitive).
          schema: {type: string}
          example: TX
        - name: zip
          in: query
          required: false
          schema: {type: string}
          example: '60603'
        - name: service
          in: query
          required: false
          schema: {$ref: '#/components/schemas/BankOperationServices'}
        - name: page
          in: query
          required: false
          description: Zero-based page index.
          schema: {type: integer, minimum: 0, default: 0}
        - name: size
          in: query
          required: false
          description: Page size.
          schema: {type: integer, minimum: 1, default: 20}
        - name: sort
          in: query
          required: false
          description: >-
            `property[,asc|desc]`; repeat the parameter for multiple sort keys. Allowed
            properties: `name`, `locationType`, `city`, `state`. Anything else returns `400`.
          style: form
          explode: true
          schema:
            type: array
            items: {type: string}
            default: [name]
          example: ['city,asc']
      responses:
        '200':
          description: One page of matching locations (empty `content` if none match).
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BankLocationPage'}
              example:
                content:
                  - id: 3
                    name: Houston Galleria Branch
                    bankAddress:
                      addressLine1: 5085 Westheimer Rd
                      addressLine2: null
                      city: Houston
                      state: TX
                      zip: '77056'
                      country: USA
                    locationType: BOTH
                    opensAt: '08:00:00'
                    closesAt: '16:00:00'
                    timeZone: America/Chicago
                    phoneNumber: (713) 555-0103
                    services: [BANKING, SAFE_DEPOSIT_LOCKER, FOREIGN_EXCHANGE, ATM_CASH_WITHDRAWAL, ATM_DEPOSIT]
                pageable:
                  pageNumber: 0
                  pageSize: 20
                  sort: {empty: false, sorted: true, unsorted: false}
                  offset: 0
                  paged: true
                  unpaged: false
                totalElements: 1
                totalPages: 1
                last: true
                first: true
                size: 20
                number: 0
                sort: {empty: false, sorted: true, unsorted: false}
                numberOfElements: 1
                empty: false
        '400':
          $ref: '#/components/responses/BadRequestText'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /v1/api/locations/{id}:
    get:
      tags: [Locations]
      operationId: getBankLocation
      summary: Get one bank location by id
      description: |
        Returns a single office/ATM. Ids are the `id` values returned by the location search.
        A non-numeric id is rejected with Spring Boot's default JSON error (`400`).
      parameters:
        - $ref: '#/components/parameters/LocationIdPath'
      responses:
        '200':
          description: The location.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/BankLocation'}
              example:
                id: 4
                name: San Antonio Riverwalk ATM
                bankAddress:
                  addressLine1: 100 E Commerce St
                  addressLine2: null
                  city: San Antonio
                  state: TX
                  zip: '78205'
                  country: USA
                locationType: ATM
                opensAt: null
                closesAt: null
                timeZone: America/Chicago
                phoneNumber: null
                services: [ATM_CASH_WITHDRAWAL, ATM_DEPOSIT]
        '400':
          $ref: '#/components/responses/BadRequestJson'
        '404':
          $ref: '#/components/responses/LocationNotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    CustomerId:
      type: apiKey
      in: header
      name: X-Customer-Id
      description: |
        Customer identifier used solely for per-customer daily rate limiting - **not**
        authentication; any non-blank value is accepted. The header name is configurable
        (`banking.rate-limit.customer-header-name`) and the check can be turned off with
        `banking.rate-limit.enabled: false`.
  parameters:
...
    LocationIdPath:
      name: id
      in: path
      required: true
      description: Bank location id.
      schema: {type: integer, format: int64}
      example: 4
...
  responses:
    BadRequestText:
      description: A business rule was violated. Plain-text message.
      content:
        text/plain:
          schema: {type: string}
          examples:
            dateRange:
              value: createdFrom (2026-05-01) must not be after createdTo (2026-01-01)
            months:
              value: months must be positive, got 0
            sort:
              value: 'Unsupported sort property: balance (supported: createdDate, closedDate, accountStatus, accountNumber)'
            locationSort:
              value: 'Unsupported sort property: phoneNumber (supported: name, locationType, city, state)'
    BadRequestJson:
      description: Request body/parameter failed validation. Spring Boot default JSON error body.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SpringErrorResponse'}
...
    LocationNotFound:
      description: No bank location with that id. Plain-text body.
      content:
        text/plain:
          schema: {type: string}
          example: 'Bank location not found: 999'
    TooManyRequests:
      description: >-
        The calling customer's daily quota (`banking.rate-limit.requests-per-day`) is
        exhausted. Written by the gateway filter as plain text. (A missing/blank customer-id
        header is rejected by the same filter with `400` and a plain-text message - see
        *Gateway behavior*.)
      headers:
        X-BTID:
          $ref: '#/components/headers/X-BTID'
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
      content:
        text/plain:
          schema: {type: string}
          example: 'Daily request limit exceeded for customer cust-42: max 1000 requests per day'
  headers:
    X-BTID:
      description: Business transaction id (UUID) for this request; appears on every server log line for it.
      schema: {type: string, format: uuid}
    X-RateLimit-Limit:
      description: Requests per day allowed for the calling customer.
      schema: {type: integer}
    X-RateLimit-Remaining:
      description: Requests remaining today for the calling customer.
      schema: {type: integer}
  schemas:
...
    LocationType:
      type: string
      description: >-
        `OFFICE` is a staffed branch, `ATM` a stand-alone ATM, `BOTH` a branch with an ATM on site.
      enum: [OFFICE, ATM, BOTH]
    BankOperationServices:
      type: string
      description: >-
        An operation a location serves. `ATM_*` values apply to `ATM`/`BOTH` locations; the rest
        to `OFFICE`/`BOTH`.
      enum: [BANKING, SAFE_DEPOSIT_LOCKER, LOANS_MORTGAGES, NOTARY, WIRE_TRANSFER, FOREIGN_EXCHANGE, ATM_CASH_WITHDRAWAL, ATM_DEPOSIT]
    BankAddress:
      type: object
      properties:
        addressLine1: {type: string, example: 5085 Westheimer Rd}
        addressLine2: {type: string, nullable: true, example: Suite 100}
        city: {type: string, example: Houston}
        state: {type: string, example: TX}
        zip: {type: string, example: '77056'}
        country: {type: string, default: USA, example: USA}
    BankLocation:
      type: object
      properties:
        id: {type: integer, format: int64}
        name: {type: string, example: Houston Galleria Branch}
        bankAddress: {$ref: '#/components/schemas/BankAddress'}
        locationType: {$ref: '#/components/schemas/LocationType'}
        opensAt:
          type: string
          format: time
          nullable: true
          description: Office opening time (`HH:mm:ss`, wall clock in `timeZone`). Null for `ATM`.
          example: '08:00:00'
        closesAt:
          type: string
          format: time
          nullable: true
          description: Office closing time. Null for `ATM`.
          example: '16:00:00'
        timeZone:
          type: string
          description: IANA zone id for the hours; `America/Chicago` is Central time and follows daylight saving.
          example: America/Chicago
        phoneNumber:
          type: string
          nullable: true
          description: Office phone. Null for `ATM`.
          example: (713) 555-0103
        services:
          type: array
          description: Operations served, in enum order.
          items: {$ref: '#/components/schemas/BankOperationServices'}
    SortInfo:
      type: object
      properties:
        empty: {type: boolean}
        sorted: {type: boolean}
        unsorted: {type: boolean}
    PageableInfo:
      type: object
      properties:
        pageNumber: {type: integer}
        pageSize: {type: integer}
        sort: {$ref: '#/components/schemas/SortInfo'}
        offset: {type: integer, format: int64}
        paged: {type: boolean}
        unpaged: {type: boolean}
...
    BankLocationPage:
      type: object
      description: Standard Spring Data `Page` serialization.
      properties:
        content:
          type: array
          items: {$ref: '#/components/schemas/BankLocation'}
        pageable: {$ref: '#/components/schemas/PageableInfo'}
        totalElements: {type: integer, format: int64}
        totalPages: {type: integer}
        last: {type: boolean}
        first: {type: boolean}
        size: {type: integer}
        number: {type: integer, description: Zero-based page index.}
        sort: {$ref: '#/components/schemas/SortInfo'}
        numberOfElements: {type: integer}
        empty: {type: boolean}
...
    SpringErrorResponse:
      type: object
      description: >-
        Spring Boot's default error body, returned when request validation fails or a
        parameter can't be parsed. It carries no field-level detail by default.
      properties:
        timestamp: {type: string, format: date-time}
        status: {type: integer, example: 400}
        error: {type: string, example: Bad Request}
        path: {type: string, example: /brite/v1/api/accounts/newaccount}